pyproject.toml
[project]
name = "compakt"
version = "0.1.0"
description = "Document summarization library using LLM-powered strategies with hierarchical structure preservation"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dependency-injector>=4.48.3",
    "docling>=2.79.0",
    "langchain>=1.2.12",
    "langchain-openai>=1.1.11",
    "langchain-text-splitters>=1.1.1",
    "markdown-it-py>=4.0.0",
    "pydantic>=2.12.5",
    "pymupdf>=1.27.2",
    "pymupdf4llm>=1.27.2.1",
    "python-dotenv>=1.2.2",
    "rapidfuzz>=3.14.3",
    "tiktoken>=0.12.0",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "azure-storage-blob>=12.28.0",
    "azure-storage-queue>=12.12.0",
    "fastapi>=0.135.1",
    "ipykernel>=7.2.0",
    "uvicorn>=0.42.0",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
